refact: Separa logica das rotas de videos para a classe VideoController em /src/controllers
- Move logica de negocios das rotas para controllers
- Mantém as rotas focadas na definição do endpoint e delegação dos controllers
- Melhora organização do código e manutenabilidade
