+++ src/routes/video.route.ts
@@ -1,7 +1,5 @@
-import express, { Request, Response, NextFunction } from "express";
-import { VideoService } from "../services/video.service";
-import { z } from "zod";
-import { CustomError } from "../types/CustomError";
+import express from "express";
+import { VideoController } from "../controllers/video.controller";
 
 export const videoRouter = express.Router();
 
@@ -35,41 +33,7 @@ export const videoRouter = express.Router();
  *  "expires_hint_hours": 12
  * }
  */
-videoRouter.post("/api/videos/metadados/client/:clientId/venue/:venueId", async (req: Request, res: Response) => {
-  try {
-    const { clientId, venueId } = req.params;
-
-    console.log("New video metadata received for client:", clientId, "venue:", venueId);
-
-    const bodySchema = z.object({
-      venue_id: z.string().uuid(),
-      captured_at: z.string(),
-
-      sha256: z.string().regex(/^[a-f0-9]{64}$/i),
-    });
-
-    // Validate Body
-    const parsed = bodySchema.safeParse(req.body);
-    if (!parsed.success) {
-      res.status(400).json({ error: "Invalid body", details: parsed.error.flatten() });
-      return;
-    }
-    const { captured_at, sha256 } = parsed.data;
-
-    const clipData = await VideoService.createSignedUrlVideo({
-      captured_at,
-      sha256,
-      venue_id: venueId,
-      client_id: clientId,
-    });
-
-    res.status(201).json(clipData);
-  } catch (error) {
-    console.error("Error processing video metadata:", error);
-    res.status(500).json({ error: "Internal server error." });
-    return;
-  }
-});
+videoRouter.post("/api/videos/metadados/client/:clientId/venue/:venueId", VideoController.createVideoMetadata);
 
 /**
  * POST /api/videos/:videoId/uploaded
@@ -92,42 +56,7 @@ videoRouter.post("/api/videos/metadados/client/:clientId/venue/:venueId", async 
  *   "status": "uploaded" | "uploaded_temp"
  * }
  */
-videoRouter.post("/api/videos/:videoId/uploaded", async (req: Request, res: Response) => {
-  try {
-    const { videoId } = req.params;
-
-    // Validate body
-    const bodySchema = z.object({
-      size_bytes: z.number().int().nonnegative(),
-      sha256: z.string().regex(/^[a-f0-9]{64}$/i),
-      etag: z.string().min(1).optional(),
-    });
-
-    const parsed = bodySchema.safeParse(req.body);
-    if (!parsed.success) {
-      res.status(400).json({ error: "Invalid body", details: parsed.error.flatten() });
-      return;
-    }
-
-    const { size_bytes, sha256, etag } = parsed.data;
-
-    const result = await VideoService.finalizeUpload({
-      videoId,
-      size_bytes,
-      sha256,
-      etag,
-    });
-
-    res.json(result);
-  } catch (error: any) {
-    if (error instanceof CustomError) {
-      res.status(error.statusCode).json({ error: error.message });
-      return;
-    }
-    console.error("Error finalizing uploaded video:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
+videoRouter.post("/api/videos/:videoId/uploaded", VideoController.finalizeVideoUpload);
 
 /**
  * GET /api/videos/list
@@ -158,60 +87,7 @@ videoRouter.post("/api/videos/:videoId/uploaded", async (req: Request, res: Resp
  *   nextOffset: number
  * }
  */
-videoRouter.get("/api/videos/list", async (req: Request, res: Response) => {
-  try {
-    // Parse and validate query params
-    const prefixRaw = typeof req.query.prefix === "string" ? req.query.prefix : "";
-
-    // Sanitize prefix
-    let prefix = "";
-    try {
-      const decoded = decodeURIComponent(prefixRaw || "");
-      if (decoded.includes("..")) {
-        res.status(400).json({ error: "Invalid prefix" });
-        return;
-      }
-      prefix = decoded.replace(/\/{2,}/g, "/").replace(/^\/+|\/+$/g, "");
-    } catch {
-      res.status(400).json({ error: "Invalid prefix" });
-      return;
-    }
-
-    const limit = Math.min(
-      100,
-      Math.max(1, Number.isFinite(+req.query.limit!) ? parseInt(req.query.limit as string, 10) : 100)
-    );
-
-    const offset = Math.max(0, Number.isFinite(+req.query.offset!) ? parseInt(req.query.offset as string, 10) : 0);
-
-    const order: "asc" | "desc" = req.query.order === "asc" ? "asc" : "desc";
-
-    const result = await VideoService.listVideos({
-      prefix,
-      limit,
-      offset,
-      order,
-    });
-
-    // Short cache for list metadata (not for signed URLs)
-    res.setHeader("Cache-Control", "private, max-age=15");
-    res.json({
-      bucket: result.files[0]?.bucket || "",
-      prefix,
-      count: result.count,
-      files: result.files,
-      hasMore: result.hasMore,
-      nextOffset: result.nextOffset,
-    });
-  } catch (error: any) {
-    if (error instanceof CustomError) {
-      res.status(error.statusCode).json({ error: error.message });
-      return;
-    }
-    console.error("Error listing videos:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
+videoRouter.get("/api/videos/list", VideoController.listVideos);
 
 /**
  * GET /api/videos/sign
@@ -230,36 +106,7 @@ videoRouter.get("/api/videos/list", async (req: Request, res: Response) => {
  *   url: string | null
  * }
  */
-videoRouter.get("/api/videos/sign", async (req: Request, res: Response) => {
-  try {
-    const path = typeof req.query.path === "string" ? req.query.path : "";
-    const kind = req.query.kind === "download" ? "download" : "preview";
-    const ttl = Math.min(86400, Math.max(60, Number.parseInt(String(req.query.ttl ?? "3600"), 10) || 3600));
-
-    // Validate path
-    if (!path || path.includes("..")) {
-      res.status(400).json({ error: "Invalid path" });
-      return;
-    }
-
-    const result = await VideoService.signUrl({
-      path,
-      kind,
-      ttl,
-    });
-
-    // Short cache for the JSON response (not the actual video)
-    res.setHeader("Cache-Control", "private, max-age=5");
-    res.json({ url: result.url });
-  } catch (error: any) {
-    if (error instanceof CustomError) {
-      res.status(error.statusCode).json({ error: error.message });
-      return;
-    }
-    console.error("Error signing URL:", error);
-    res.status(500).json({ error: "Internal server error" });
-  }
-});
+videoRouter.get("/api/videos/sign", VideoController.signVideoUrl);
 
 /**
  * GET /videos-clips
@@ -281,23 +128,4 @@ videoRouter.get("/api/videos/sign", async (req: Request, res: Response) => {
  *   }>
  * }
  */
-videoRouter.get("/videos-clips", async (req: Request, res: Response) => {
-  try {
-    const venueId = req.query.venueId as string;
-
-    if (!venueId) {
-      res.status(400).json({ error: "venueId is required" });
-      return;
-    }
-
-    const result = await VideoService.getClipsByVenue(venueId);
-    res.json(result);
-  } catch (error: any) {
-    if (error instanceof CustomError) {
-      res.status(error.statusCode).json({ error: error.message });
-      return;
-    }
-    console.error("Error fetching clips:", error);
-    res.status(500).json({ error: "Erro ao buscar clipes gerados" });
-  }
-});
+videoRouter.get("/videos-clips", VideoController.getClipsByVenue);
